requirements.txt
requests==2.31.0
feedparser==6.0.10
pytube==15.0.0
yt-dlp==2023.12.30
openai-whisper==20231117
moviepy==1.0.3
opencv-python==4.8.1.78
numpy==1.24.3
selenium==4.15.2
schedule==1.2.0
openai==0.28.1
python-dotenv==1.0.0
Pillow==10.0.1 